Move impl to transformers.rs
